change main example, fix typo

+++ polyomino.jl
@@ -369,7 +369,7 @@ function minRooks(p::Polyomino)
             end
 
             if (sum == guarded)  # if sum is "1111...111", so all tiles are guarded
-                rooks = Pair{Int64, Int64}[]
+                rooks = Pair{Int128, Int128}[]
                 for h in j
                     push!(rooks, tileIdRev[bitsetOrder[h]])  # retranslate bitset to original tile
                 end
